Log refactoring.

+++ ParallelRoadTool/Utils/Log.cs
@@ -1,11 +1,11 @@
-﻿namespace CSUtil.Commons
-{
-    using System;
-    using System.Diagnostics;
-    using System.IO;
-    using System.Threading;
-    using UnityEngine;
+﻿using System;
+using System.Diagnostics;
+using System.IO;
+using System.Threading;
+using UnityEngine;
 
+namespace CSUtil.Commons
+{
 #if !DEBUG
 #if TRACE
 #error TRACE is defined outside of a DEBUG build, please remove
@@ -16,35 +16,23 @@
     {
         private static readonly object LogLock = new object();
 
-        // TODO refactor log filename to configuration
-        private static readonly string LogFilename
-            = Path.Combine(Application.dataPath, "PRT.log");
+        private static readonly string LogFilename = Path.Combine(Application.dataPath, "PRT.log");
 
-        private enum LogLevel
-        {
-            Trace,
-            Debug,
-            Info,
-            Warning,
-            Error
-        }
-
-        private static Stopwatch _sw = Stopwatch.StartNew();
+        private static readonly Stopwatch _sw = Stopwatch.StartNew();
 
         static Log()
         {
             try
             {
-                if (File.Exists(LogFilename))
-                {
-                    File.Delete(LogFilename);
-                }
+                if (File.Exists(LogFilename)) File.Delete(LogFilename);
             }
-            catch (Exception) { }
+            catch (Exception)
+            {
+            }
         }
 
         /// <summary>
-        /// Will log only if debug mode
+        ///     Will log only if debug mode
         /// </summary>
         /// <param name="s">The text</param>
         [Conditional("DEBUG")]
@@ -54,7 +42,7 @@
         }
 
         /// <summary>
-        /// Will log only if debug mode, the string is prepared using string.Format
+        ///     Will log only if debug mode, the string is prepared using string.Format
         /// </summary>
         /// <param name="format">The text</param>
         [Conditional("DEBUG")]
@@ -64,9 +52,9 @@
         }
 
         /// <summary>
-        /// Will log only if debug mode is enabled and the condition is true
-        /// NOTE: If a lambda contains values from `out` and `ref` scope args,
-        /// then you can not use a lambda, instead use `if (cond) { Log._Debug }`
+        ///     Will log only if debug mode is enabled and the condition is true
+        ///     NOTE: If a lambda contains values from `out` and `ref` scope args,
+        ///     then you can not use a lambda, instead use `if (cond) { Log._Debug }`
         /// </summary>
         /// <param name="cond">The condition</param>
         /// <param name="s">The function which returns text to log</param>
@@ -74,10 +62,7 @@
         [Conditional("DEBUG")]
         public static void _DebugIf(bool cond, Func<string> s)
         {
-            if (cond)
-            {
-                LogToFile(s(), LogLevel.Debug);
-            }
+            if (cond) LogToFile(s(), LogLevel.Debug);
         }
 
         [Conditional("TRACE")]
@@ -92,7 +77,7 @@
         }
 
         /// <summary>
-        /// Will log a warning only if debug mode
+        ///     Will log a warning only if debug mode
         /// </summary>
         /// <param name="s">The text</param>
         [Conditional("DEBUG")]
@@ -102,19 +87,16 @@
         }
 
         /// <summary>
-        /// Log a warning only in debug mode if cond is true
-        /// NOTE: If a lambda contains values from `out` and `ref` scope args,
-        /// then you can not use a lambda, instead use `if (cond) { Log._DebugOnlyWarning }`
+        ///     Log a warning only in debug mode if cond is true
+        ///     NOTE: If a lambda contains values from `out` and `ref` scope args,
+        ///     then you can not use a lambda, instead use `if (cond) { Log._DebugOnlyWarning }`
         /// </summary>
         /// <param name="cond">The condition</param>
         /// <param name="s">The function which returns text to log</param>
         [Conditional("DEBUG")]
         public static void _DebugOnlyWarningIf(bool cond, Func<string> s)
         {
-            if (cond)
-            {
-                LogToFile(s(), LogLevel.Warning);
-            }
+            if (cond) LogToFile(s(), LogLevel.Warning);
         }
 
         public static void Warning(string s)
@@ -128,23 +110,26 @@
         }
 
         /// <summary>
-        /// Log a warning only if cond is true
-        /// NOTE: If a lambda contains values from `out` and `ref` scope args,
-        /// then you can not use a lambda, instead use `if (cond) { Log.Warning }`
+        ///     Log a warning only if cond is true
+        ///     NOTE: If a lambda contains values from `out` and `ref` scope args,
+        ///     then you can not use a lambda, instead use `if (cond) { Log.Warning }`
         /// </summary>
         /// <param name="cond">The condition</param>
         /// <param name="s">The function which returns text to log</param>
         public static void WarningIf(bool cond, Func<string> s)
         {
-            if (cond)
-            {
-                LogToFile(s(), LogLevel.Warning);
-            }
+            if (cond) LogToFile(s(), LogLevel.Warning);
         }
 
         public static void Error(string s)
         {
             LogToFile(s, LogLevel.Error);
+        }
+
+        public static void Exception(Exception e)
+        {
+            LogToFile($"Exception detected, message is: {e.Message}", LogLevel.Exception);
+            LogToFile(e.StackTrace, LogLevel.Exception);
         }
 
         public static void ErrorFormat(string format, params object[] args)
@@ -153,7 +138,7 @@
         }
 
         /// <summary>
-        /// Log error only in debug mode
+        ///     Log error only in debug mode
         /// </summary>
         /// <param name="s">The text</param>
         [Conditional("DEBUG")]
@@ -163,7 +148,7 @@
         }
 
         /// <summary>
-        /// Writes an Error message about something not implemented. Debug only.
+        ///     Writes an Error message about something not implemented. Debug only.
         /// </summary>
         /// <param name="what">The hint about what is not implemented</param>
         [Conditional("DEBUG")]
@@ -178,18 +163,18 @@
             {
                 Monitor.Enter(LogLock);
 
-                using (StreamWriter w = File.AppendText(LogFilename))
+                using (var w = File.AppendText(LogFilename))
                 {
-                    long secs = _sw.ElapsedTicks / Stopwatch.Frequency;
-                    long fraction = _sw.ElapsedTicks % Stopwatch.Frequency;
+                    var secs = _sw.ElapsedTicks / Stopwatch.Frequency;
+                    var fraction = _sw.ElapsedTicks % Stopwatch.Frequency;
                     w.WriteLine(
-                        $"{level.ToString()} " +
+                        $"[{level}] " +
                         $"{secs:n0}.{fraction:D7}: " +
                         $"{log}");
 
                     if (level == LogLevel.Warning || level == LogLevel.Error)
                     {
-                        w.WriteLine((new System.Diagnostics.StackTrace()).ToString());
+                        w.WriteLine(new StackTrace().ToString());
                         w.WriteLine();
                     }
                 }
@@ -198,6 +183,16 @@
             {
                 Monitor.Exit(LogLock);
             }
+        }
+
+        private enum LogLevel
+        {
+            Trace,
+            Debug,
+            Info,
+            Warning,
+            Error,
+            Exception
         }
     }
 }
